Fix crc comparison in multidex installation.
The bug was caused by a comparison of Long instances instead of long
values and was leading to the library always detecting that the
secondary dex files needed a fresh exctraction.

bug 12776578

+++ library/src/android/support/multidex/MultiDexExtractor.java
@@ -191,7 +191,7 @@ final class MultiDexExtractor {
 
         // We know the length of storedDexCrcs and dexCrcs are the same.
         for (int i = 0; i < storedDexCrcs.size(); i++) {
-            if (storedDexCrcs.get(i) != dexCrcs.get(i)) {
+            if (storedDexCrcs.get(i).longValue() != dexCrcs.get(i).longValue()) {
                 return true;
             }
         }
